requirements.txt
python-dotenv
openai
pyupbit
pandas
ta